Floor call duration to 0 seconds
Sometimes the Gaussian distribution results in negative call
durations, which is not possible. In this case, the duration is
set to 0 seconds.

There is likely a better way to do this that weights the low
likelihood of implausibly short contacts with some more stats
math but that will be for another day.